requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
starlette==0.38.5
uvicorn==0.30.6
